client/requirements.txt
gcloud-aio-pubsub
gcloud-aio-run
